change algorithm to calculate assigned days for an issue.Add date-fns library

+++ src/pages/main/components/issue-table/IssueList/IssueList.js
@@ -8,9 +8,12 @@ export default function IssueList({issues, setEditIssueActive}) {
     const dispatch = useDispatch();
 
     function deleteIssueHandler(issueId) {
-        dispatch(deleteIssue(issueId));
-        dispatch(deleteDaysByIssueId(issueId));
+        if (window.confirm("Are you sure you want to delete the issue?")) {
+            dispatch(deleteIssue(issueId));
+            dispatch(deleteDaysByIssueId(issueId));
+        }
     }
+
     function editIssueHandler(issueId) {
         setEditIssueActive({isActive:true, issueIdToEdit:issueId});
     }
